+++ CHANGELOG.md
@@ -102,6 +102,8 @@
 - PR #5591 Add `__arrow_array__` protocol and raise a descriptive error message
 - PR #5601 Instantiate Table instances in `Frame._concat` to avoid `DF.insert()` overhead
 - PR #5602 Add support for concatenation of `Series` & `DataFrame` in `cudf.concat` when `axis=0`
+- PR #5603 Refactor JIT `parser.cpp`
+- PR #5643 Update `isort` to 5.0.4
 
 ## Bug Fixes
 
@@ -152,6 +154,8 @@
 - PR #5490 Fix python column view
 - PR #5608 Fix issue related to string types being represented as binary types
 - PR #5619 Fix issue related to typecasting when using a `CategoricalDtype`
+- PR #5649 Fix issue when empty Dataframe with index are passed to `cudf.concat`
+- PR #5644 Fix issue related to Dataframe init when passing in `columns`
 
 
 # cuDF 0.14.0 (Date TBD)